Adds an option to admin email users from manager
The User manager page now has a button to send the selected user ids
to the Admin emails page.

The Admin emails page will default to the selected user ids in the
csv id entry box.

When the Admin emails page is refreshed it will reset to the ids from
the Managers page.