Improve UI/UX when selecting unbooked students
The 'Select all unbooked students' checkbox toggles on and off all the unbooked students. When the checkbox is on and any of the students are toggled off, the checkbox gets toggled off. When all the unbooked students are individually toggled on, the checkbox does NOT get turned on. (We may want to consider this, though).

Also fixes a bug where students who did not allow full access were selectable via the 'select all' checkbox.