Fix exam boards for those who haven't selected one
The site would default to OCR for all users who had not yet selected
an exam board.
This commit adds a third option "Other" which functions similarly to
the exam board selection for anonymous users with a selection box to
choose the exam board on each page.
Additionally the exam board is added to the required account info
so no future users end up with exam boards as undefined.

The exam board preference was moved from UserPreferences to User and
switched from two boolean flags to a string making further exam boards
easer to add.