Integrate google maps with the events page
Add a map to the events page displaying all of the events.
Uses the @react-google-maps/api package, chosen over google-map-react
to be more actively maintained.

Note:
The state variable mapCenter prevents the map from recentering to
the starting location (Cambridge) when ever an info window is hidden.

Currently a request is made each time the events filter changes
in order only show the correct events. Ideally this only needs to be
one request which provides all event locations and to filter locally.