Upgrade format of package-lock.json to avoid constant nagging from LGTM.
This DOES NOT change any installed versions, but only marks some packages as optional, so it should be a safe upgrade. I am once again not sure if this is the nth difference between npm on macOS and Windows but hey, what can you realistically expect from the bunch of monkeys running the JS community these days?