Add a warning for users typing unclear function calls in text entry
If a user types "Tsin(theta)" this will be parsed as "T*s*i*n*theta", which
is not what the user intended. Whilst this is visible, in the fact that the
result has italic and not upright letters for the sin in the preview, this
fact is not obvious to most users. Instead add a warning underneath.
We have to allow "a" and "c" as characters before to allow "asin(x)" or
"arcsin(x)", and this seems the simplest regex to achieve this.