Rename LoggedInUser type to allow use of type guards in user check service
The LoggedInUser type was not actually a logged in user, and so you could
not make a type assertion in services/user.tsx about the return value of
`isLoggedIn(user)`. Move to PotentialUser as the type for a user that might
be logged in, and use LoggedInUser as a wrapper to a RegisteredUserDTO for
an actual user object.