Attempt to speed up page load time by preloading Exo2 fonts
This might not work, but it does seem to in Chrome and Firefox and perhaps
subjectively does seem to make a difference. Certainly the fonts are loaded
much earlier on in the network trace, and we're mostly HTTP/2 so they won't
even block other requests!
This will need updating if the fonts ever change name, etc, but Chrome at
least warns about unused preloaded resources.