Support disabling indentation for Parsons problems
It is sometimes desirable to prevent items from being indented in Parsons problems. This uses the IsaacParsonsQuestionDTO::disableIndentation flag to achieve the effect. This will require a toggle control to be added in the editor to turn the flag on and off.

Some minor changes here and there, notably changing == to ===.