Fix a bug with negative numbers in Inequality's menu
If numbers are pressed after pressing the plusminus button, the new digits are added to a negative number instead of subtracted.

This also fixes another definedness bug where 0 and undefined are considered the same thing by Javascript, thus showing the wrong thing in the number box when 0 is selected first.