Run npm dedupe and ensure we use the minified p5 to keep bundle size down
This is not the correct solution at the end of the day; we are still adding
p5 in two places, inflating the bundle by ~600kb compressed JS. But it is a
vast improvement.
I don't know why `npm dedupe` is needed again, but we should be careful and
run it and commit the results more regularly.

On the p5 front, compare the import of the graph sketcher in
IsaacGraphSketcherQuestion:
```
import {GraphSketcher, makeGraphSketcher, LineType, GraphSketcherState}
    from "isaac-graph-sketcher/dist/src/GraphSketcher";
```
to the import of Inequality:
```
import { Inequality, makeInequality } from 'inequality';
```
we should review that and see if we can make them consistent!