Change the seeded random number generator to reduce bundle size
To randomise the units shown on numeric questions, we were using
`math-random-seed`, which is a wrapper around another library built for
NodeJS that assumes the existence of a `crypto` module. Webpack will add in
the full crypto code for you if you're in a browser, but we don't need
99% of it. In fact, we don't even need pseudo random numbers for this
use-case, just something to muddle the ordering of units a little.
Replacing the library also removes the need for `bn.js` too.
This is a 0.8MB saving, or 11%.

For future reference, adding `node: false,` to your Webpack config will
prevent it from bundling NodeJS polyfills automatically and ensure this
cannot accidentally happen. But other libraries rely on the `global=window`
polyfill this does, so I'm not including that change now.