Fix errors caused by new random number library
The library seems to use the "this" keyword in a way that is no longer
 bound to the correct thing when using the
`random = new Rand(seedValue).next`
shorthand. I don't understand "this" well enough, so just use the form from
the documentation, which fixes the issue. I clearly didn't test it well
enough when I refactored it!