Ensure My Progress page graph always has sensible axis range
Billboard does very silly things when there is only one datapoint, and uses
a whole decade as the scale for the axis. Unfortunately if you have one
point and one label, the label is not underneath the point, so you need
three labels in that case. In general we do not want this additional
spacing nor odd numbering, and so we only do it in this case.
Also remove unnecessary parsing and converting of dates back to strings,
since they went from `YYYY-MM-DD` form back to `YYYY-MM-DD` form!