Show search level and topic separator only if both have displayed values
It's a mess, since displayTopicTitle can be true but the topic title empty
so we showed an unnecessary comma. Then we would have had to repeat the
condition again to decide whether to show the comma. So I moved the comma
span inside the topic title span so that it is only shown if necessary!