Add figure numbering to page fragments
This is achieved by creating a figure numbering "scope" for each page - unfortunately piggy-backing on the already overloaded TrackedRoute component.
I believe that it would be possible for a page with two page fragments to load out of order so that Figure 2 appears before Figure 1 on the final rendered page but that did not happen in my constructed page.
No examples exist in the wild so I constructed a page with two page fragments each containing a uniqu Figure and figure reference.
It certainly isn't perfect but it does the right thing in more cases than it used to.