Rewrite video ID matching to avoid lookbehind groups unsupported in Safari
This ought to be slightly more flexible than the previous Regex too, since
it ought to also support youtu.be/[video_id] examples that the previous one
did not. But it might also be more prone to break. We shall see!
Of course it is Safari that doesn't do sensible things . . .