Emergency fix for inequality menu.
The moving item's content would overflow. In particular, \sqrt goes on forever. This might be a but with katex but for now this commit patches things up until we can investigate further. The solution is not ideal but the moving item's representation also has problems with big numbers overflowing, so maybe it's time for a different approach?