Allow the compose scripts to manage both physics and CS containers
This also improves the validation of inputs initially and before they get
passed into docker-compose, which seemed sensible.
It also removes the "special-case" nature of the test database
initialisation; since the Docker Postgres image will not use the init
scripts if the data directory is blank, it ought to be safe to include the
test scheme for all non-prod databases. Doing this dramatically simplifies
the compose script.
This also restricts the types of docker-compose commands that can be run,
to prevent passing unexpected things into it. If any commands have been
missed, they can easily be added and a helpful error is printed.