Correctly name placeholder buttons for links to events and resources
I've also added links to teachcomputing.org on the logos. I've set these as "noopener" but not "noreferrer" currently, because "noopener" stops the security risk of the linked page having access to the linking page via `window.opener`. The latter stops the linked site knowing that we referred them via our site - since the site in question is teachcomputing.org I thought it would be more desireable that they could know/it doesn't matter if they can or not.