Add table view and add groups modal to set assignments page
I have moved a lot of the code and logic from MyGameboards across to this, editing to
remove filtering by completion as I guessed that it made no sense here. I feel like sorting
by number of assigned groups would be nice to take the place of sorting by completion.

I still need to add an explicit button to open the group assigment modal for each row, and
perhaps refactor out the grid and card view patterns from SetAssignments and MyGameboards
to be more maintainable.