Add filtering by subject in table view
Filtering checks if the boards contain the selected subject - another way could be to
filter similarly to levels, and allow selection of multiple subjects, checking if the board
subject set is equal to the set of selected ones. Both have upsides.

I've kept the modal, and formatted the close button to be the same as ActiveModal.
Another reason that ActiveModal doesn't work well here is that it is visually too
wide in this use case - the assign/unassign menu looks much better portrait/longways,
whereas ActiveModal seems to be for landscape/sideways content.