Remove landing page content for logged in user
Intro video, welcome text, and NCCE logos are now removed for logged in users, instead only greeting the user
in a header, keeping links to GCSE/A Level topics and events.

In line with the suggestion to remove the featured question and keep the CS journey, I've reorganised the featured
tabs so that the CS journey is shown first - the featured question is still there but is massively de-emphasised now
that it is in the second tab.