Fix differing widths of GCSE and A Level buttons on small screen sizes
I thought they should be considered equal visually, so made sure they were
always the same size on different size screens. In the previous commit,
the stacking on small screens was fine but a point could have been made that
it de-emphasised the A Level button. It is now balanced, as all things should be.