Add boolean notation dropdown to `RequiredAccountInformationModal`
Also refactored the code in line with Meurig's suggestions, and moved the boolean notation input into a
separate component similar to `StudyingCSInput`.

The keys of `BooleanNotation` are stored as an enum for easy mapping over, and the `BOOLEAN_NOTATION`
beta flag has been removed, with all occurences changes to `AUDIENCE_CONTEXT` instead.