Un-require boolean notation preference, and redesign select choices
When the user visits their account page, the default value of the dropdown is "No preference", which
maps to both flags (`MATH` and `ENG`) being set to `false`. This means the user doesn't really have to
think about this field, and you have to mess with the code to make the field invalid (as any options are
valid).