Add generate gameboard from concept button to PHY concept pages (for demo)
This is for demos. The content of this commit might have to be commented out or reverted later.

It does however highlight a weird bug where the gameboard filter doesn't update based on query
parameters when visited by a hard link (opening the filter in a new tab) or for some related reason.
They do update when the page is refreshed.