Upgrade inequality and isaac-graph-sketcher
The two packages now require p5 and lodash as peer dependencies so they are not bundled within each of them and thus imported twice into isaac-react-app. The "downside" is that p5 and lodash must now be installed manually alongside each package. This is not a problem, but npm install currently warns us of a bunch of uninstalled peer dependencies which we can safely ignore because we don't use, so the warning about p5 and lodash may get drowned.

Bottom line: don't mess with p5 and lodash from package.json.