Preserve topic exam board tab by recording it in URL hash
This was requested in ticket #204947. When the back button is pressed after leaving the CS all topics page, the exam board tab the user was on was not recorded, so the user would have to navigate back to the tab every time. This is no longer the case - the active tab is taken from a URL hash (either 'all', or a lowercase exam board name) and this hash is updated when the tabs are changed.
This could be abstracted to a wrapper component around Tab that uses the URL hash to choose the active tab, but this would clash with components that also use the hash for other functions etc. so it doesn't seem worth it yet.