Generify role checking for user objects
This modifies the role checking functions so that they work on any "user" type containing either or both of `role` and `loggedIn` attributes.
TS is clever enough to work out that, even though both `role` and `loggedIn` are optional params, if you pass an object that has neither of those params then you are probably making a mistake and it gives you a nice "types don't overlap" warning.

This change also demonstrates a nice way of defining a union type using an array of strings, so that you can define orderings (for example) on the values in the array while keeping things type-safe.