Add new login or sign up modal
This also changes how redirects after account update are handled by adding an extra argument to the `updateCurrentUser` action creator, specifically to stop the required account info modal from redirecting to My Account. It doesn't make sense that it does IMO, since it's disruptive to the user, and it messes with how this new modal redirects the user back to the question they were attempting after they sign up.

I also added an optional `noPadding` field to `ActiveModal`, to remove the padding around the `ModalBody` element for when you want the content to be flush to the edges of the modal (I needed this because the new modal has a background on half of it).