Add rel=canoncial meta tags to pages with many URL parameters
When a search engine indexes these pages, we only care about the full
versions of the pages (I think!) and not the exam board or stage specific
versions. This strips out all URL parameters and hashes and leaves just
the bare URL as "canonical".
The new element is designed to work on any page so it can be reused.