Enable 'expand' button optionally for Table and Code Snippet
Makes the expand button disabled by default, and enabled via content
editor options. For the table, being HTML in a regular Content block,
this is enabled by providing class="expandable". For the code snippet
there is a checkbox. Tabs still show the button by default for now,
this will change.