Fix some tables not rendering correctly
`getElementsByTagName` returns an `HTMLCollection` that updates to reflect DOM changes - I was iterating over a collection of `table` elements and removing them from the DOM, but *not* from the collection, so I assumed the size of that collection would stay static. However, it did not, because the collection updated to reflect the new DOM and we effectively ended up in a situation where an array was having its length decreased while being iterated over with a for loop, resulting in some elements of the collection not being processed.