Standardize the "portal element in HTML" hook pattern
This moves cloze drop-zone `div` generation into `TrustedMarkdown` (which requires trying rendering KaTeX in `TrustedMarkdown` as well - I've tried to make this as nice as I can; as a plus it does mean glossary terms in LaTeX are theorectically possible).
I also refactored `TrustedMarkdown` so that the process of how the markdown is rendered is crystal-clear - each separate process has its own function, and they are composed to create the whole process.
The hooks for rendering elements in `dangerouslySetInnerHtml` portals are now managed in a single file (`portals/utils.ts`), with two constants listing all of the hooks of this type available, and just those that accessible tables can use (for now).