Evil hack to get keyboard working on cloze questions
This takes `useKeyboardSensor` from `react-beautiful-dnd`, and adds some extra (nasty) stuff so that it behaves as expected with cloze questions. Since you can only interact with one cloze question at a time, this should work fine in theory, and is certainly better than the non-functioning keyboard interaction we had already.
It might break if the question rerenders underneath it, but it should break nicely and just cancel any drag in progress.