Move updating of quiz attempts into reducer, rather than selector
This moves logic that updates the current users quiz attempt with their current best attempt from the quiz attempt selector, into the quiz attempt reducer (so it now gets updated when a "question attempt" redux action is fired)

This also refactors the `pageSettings` stuff that was previously passed down the tree of the progress pages. I changed it to be a React context instead, to have including it in the props of loads of components.