Fix two WebSocket bugs with missed function params
This code originally came from the AngularJS world, where checking
the root scope for the user was done inside each function and so
they had no arguments. This meant that passing setTimeout just the
function was enough. When the user argument was added to check there
was a user, it was missing from all the setTimeout calls.

This caused two issues:
 - After an error, WebSockets would not reconnect because the userId
   parameter was never provided to openNotificationSocket.
 - Rather than sending heartbeats, we requested a full userSnapshot
   every 60s because the userSnapshot argument was never passed to
   checkForWebSocket.

The fix is simple; return to checking the state inside each function
where it matters. Slightly hacky, but safe for this use-case.
Also smooth now-fixed reconnect attempts over an additional 30s period
due to higher traffic volumes.