Add book board and gameboard builder buttons to assignment schedule
Requires redirecting user back to assignment setting page they originally pressed these buttons from - I save the path to redirect to in local storage and default to "/set_assignments" if it's missing.
It won't be possible for a user to accidentally be redirected to the assignment scheudle unless they've visited it already in that session and take a weird path through the app.