Fix bug on all browers by forcing root div to have `scrollLeft = 0` if it scrolls
This plays nicely with the fix for newer browsers in the previous commit, and is fairly lightweight and self-contained.
If the browser doesn't have `overflow: clip` support, there is a small discrepancy between where an item is dropped and where its drop animation indicates it will be dropped, but this only happens where the bug would occur previously and is a big improvement.