Restore core-js but only for used features
Apparently some users are using browsers over 3 years out of date, but
still "modern" enough that we cannot use the Nginx User-Agent detection
to block them. The polyfills are hence just for the modern features we
use that are causing problems, and is much smaller than including all of
core-js.
This replaces the additional string.prototype.matchall library with the
same feature already in core-js, and removes Element#closest polyfills
which are not necessary outside of IE.