Allow p5 JS chunk to be lazy loaded separately
RoutesPhy had a direct, non-lazy, path to loading p5 and Inequality.
All the other paths were lazy, and so swapping that one for a lazy path
too allowed Webpack to do its magic.
The only other change is a minor cleanup of an unused import that may
also have triggered a non-lazy load.
This dramatically reduces the size of the main bundle!