Fix the user consistency modal notification
Swap setImmediate(...) with setTimeout(...) as the former is no longer polyfilled.
The two functions are not strictly equivalent as nested setTimeouts, once nested more than five deep, are forced to wait at least 4 milliseconds.
This difference should be fine for our usecase even though we are nesting the calls.

This was the only call to setImmediate(...) in the codebase.