Work around weird KaTeX phantom element behaviour
Most noticeable in our chemistry content, there are mysterious hidden
X characters used for separating and spacing that show up when copying
and pasting text.
In https://github.com/KaTeX/KaTeX/issues/3668, a suggested fix was to
use visibility: hidden, which seems to be the correct way of having a
spacing character. Until that issue is resolved upstream, hack this
behaviour in by replacing the current transparent method style text.
This fixes much of the weirdness of copy and paste, although there
are still weird newlines appearing that I cannot track down.