Use error.message not error.stack for reporting
The error message contains the key issue, and the stack trace in the
production build is rarely if ever helpful. We can still ask for the
trace from Dev Tools if it proves really necessary.
Also use the correct encodeURIComponent for encoding a URI component,
but this didn't fix the newline issue we'd seen.