Render error message rather than error object which causes errors
Global error state in redux is and should be avoided generally (because some errors end up being shown on incorrect pages).
The problem here was in the naming of the variable as errorMessage whereas in fact the variable held the error object.
React refuses to render objects, throws an exception and then hits our app error boundary.