Add scroll-to-top button to Ada
ScrollToTop component added to the IsaacApp component. It remains at the bottom of the screen (out of site) until the user scrolls beyond the top of the main content at which point it sticks to the bottom right of the screen.

Currently added only to the Ada site but removing the conditional will add it to the Physics site as well. The scroll-button.scss will also need to be imported into the relevant isaac.scss file.