Display the pre-selected year despite of year range
If a due date was in a prior year attempting to change the date in a
DateInput would change the year to the current year as the prior year
was not an option.
This would not update any of logic and so relevent error messages would
display whilst visually the messages made no sense.

The invalid year is not presented as an option and once a valid
selection is made there is no way for the user to reselect the original
date. A user can exit out of the modal or page to maintain the prior
date.