Upgrade billboard.js to 3.10.3
Minor changes to the chart generation.

The overlapping legend text on hidden charts is now fixed with a new
`render` property. This, however, means that the charts will only
display once flushed. If the flush is run on load then the charts have
not been generated by billboard and so no change occurs. A `useEffect`
on the `onActiveTabChange` (which is `flush` for charts) runs the flush
when these have been generated thus displaying the charts on load.

A better solution likely exists.