Update dependencies and Node build version
This version of Babel requires Node as least 18.12, so this upgrades the
Actions runner to the latest LTS.
@babel/plugin-proposal-class-properties hasn't been a proposal in 4
years, so use the actual transform package now.
Axios is now 1.0, but has changed the type of the response data and so
I have used an any as a quick fix.

There are many changes here, I have tested what I could.