Upgrade Jest to latest version
We were using the uuid package, which is problematic in newer versions
of Jest. It has been fixed in the latest version, but we were using it
without declaring a dependency (it happened to be a transitive dep of a
dev dependency!). Added the latest version, upgraded webpack-dev-server
which transitively used it.
We now need to explicitly add jest-environment-jsdom for newer versions
of Jest.