Restore prominent upgrade role link to My Account
Annoyingly the Ada designers want bolded links, which means a bunch of
siteSpecifics. Also remove the existing text next to the context picker
since it was only shown to students but only helpful to teachers.
Role names are now lowercase, but these are only used by the
AccountTypeMessage component anyhow.