Stop double decoding the URI
The `decodeURI` was called on the query string as it was encoded when it
was pushed to the URI. This was to prevent issues with `/` and `%` that
may occur.

However, the `useQueryParams` hook already provided a decoded URI. This
resulted in `%` being encoded as `%25` then decoded to `%` which failed
the second decode.