Display warning banner when in revision mode
If a user is in revision mode (thus their previous attempts are hidden)
and are attempting a question where some of the question parts have
previous attempts a banner is shown at the top of the question
explaining why their attempts have been hidden.

There is a link to the beta feature page to turn the revision mode off
if the student wishes to see their previous answer.