Update search string URL component only on enter
To prevent slowing down the search string filter the URL component and
search results are only updated on enter. This is how the `Search` page
operates.

A better method of achieving this may be possible. The `Search` page
mechanism seems too different from the current `QuestionFinder` page to
try and use that version.