Focus the response banner after submission
After submission the focus on the page resets to `body` which means
using a screenreader (or tabbing) takes the user to the top of the page.

As the submit button disappears after a correct submission the response
banner is focused instead. This way the user can continue from the
question answer. This does require making the `div` tabbable,
which should ideally mean that element is responsive (which it is not),
as otherwise it cannot be focused on.