Allow deploy script on the standby server
The deploy script currently expects previously running containers to
exist. It both queries these containers and brings them down.

The standby server has no containers running and therefore the script
fails when attempted on the standby server. This PR skips commands that
use the old containers when it detects they are not running and adjust
commands as necessary.

Support for staging, test, and dev have not been added as these are
unlikely to be required on the standby server.