Use dropdown menus for cloze questions on mobile
When the screen is small (e.g. a mobile device) the drop zones are
converted to dropdown menus with the available items. The item tray is
removed.

If there is no replacement then selecting an item removes it from other
dropdowns. A blank option is added to allow for resetting individual
zones.

This requires a new handler function in the IsaacClozeQuestion component
that is passed to the InlineDropZone rendering component through the
DropRegionContext. This is an altered version of the onDragEnd handler.