Show "Core" and "Advanced" on off spec content
If a topic or concept is not on a users specification display the tag as
"core" or "advanced" based on prior tagging. This prefers "core" over
"advanced" and only shows "advanced" if "core" is not present.

If a user has not set their preferences every tag is displayed as "core"
or "advanced". To accomodate this, if a user has not set their preferences
, on Ada, they have a new "none" stage and examboard to distinguish from
deliberately selected "all".