Fix use-before-initialisation error in symbolic qs
The bug is tricky. The function and useState are defined using the
currentAttemptValue name before it is defined by the let. Without the
let, this would refer to window.currentAttemptValue, but as soon as the
let occurs, this reserves the name in the local scope but does not
initialise the variable.
If the useState calls the currentAttemptPythonExpression function before
the variable is initialised inside the if below, then you get a
reference error.

However, all of this has been true for 4 years. Why only now is it
suddenly breaking? :/