Set email preference defaults on account creation for Ada
In the old signup flow, the required account info modal would pop up
incessantly if this was left undefined (although could still be skipped).
In the new flow, the preferences stage of the flow is a lot easier to
skip once and forget about, so it's best to set this upon initial
account creation and update it later.