Ignore default exam board if set by query param
There is an order of presedence that we follow when deciding which exam board to use for a user's viewing context. Query params have the heighest presedence as the user should see the same page that the link sender wanted to show you.

We must ignore this case when the query params are the default params. The query params are updated to reflect the viewing context so that people's links coppied from the URL bar are correct. However, if we follow a link with no query params as a signed in user, for the initial render we don't know that you're signed in or not so give you the default values and set it in the query params. When we then receive the signed in request we would like the account preferences to override it.

A better solution exists but will need to wait until the next release.