Restore old question finder URL and behaviours
The changes in 0e50cf17e0bdd61f53adae65c20facd48cc27f77 and in
3dfccb91d52022e4fbd82f6fd29a957455b7cb47 removed the old gameboard
filter version of the question finder entirely, whilst also breaking
several other pages that relied on its functionality.
This restores the old URL, and changes the broken pages to use the old
version until that is finally removed.