Avoid loading both CJS and ESM versions of RTKQuery
Here the problem was more subtle; we were sometimes loading from /dist/
which is the internal namespace that is not supposed to be used.
Mostly this was a case of removing the /dist/, but for some specialised
cases we do still need the internal type. So swap to a type import to
avoid a runtime dependency.