Stop setting viewing context defaults on registration as not saved in DB
The code, for the previous couple of months, has tried to set the default viewing context during registration but the back-end code ignores it. We actually do not want to be setting the default viewing context at this step in the registration because we want to continue to display the context picker to users who have not explicitly set them.