Update q finder filter from account settings
The previous code set the stage and exam board settings from the user's viewing context. Although the naming is a little complicated, the user's viewing context is not the same as the user's account level registered contexts. It is those account level registered contexts that we are interested in for pre-populating filters.
If the user comes to the question finder with query params specifying the stages and exambaords filter settings they take presedence over account settings.