Default exam spec tabs to user context
On Ada: if an exam specification page has stage/exam board tab options that match a user's context, default to those tabs when the page is loaded.

This works best for users with a single context set; if a user (teacher or above) has multiple contexts set, the tabs will default to the last match found.