Add admin search postcode validation
Add a regex to verify that a postcode is of a sensible format in admin user search before sending it to the API. The regex should match complete postcodes (e.g. CB3 0FD) and outcodes (e.g. CB3).

This check doesn't verify that a postcode actually exists, so a server error can still be thrown in some cases, but this should reduce the chances of this happening.