Replace or remove references to gameboard filter
Since the filter is gone, there is no need to allow the "filter=true"
URL parameter to work any more.
We may want to re-use the code from extractFilterQueryString if we ever
decide to add links from gameboards to the new question finder, but I
have removed it for now since it was unused and doing so allows removing
the library too.