Modify news/event pod styling
Lowers the container breakpoint slightly to fix pod behaviour around 790px on News/Event pages (pod images were previously using the full screen width at this size).

Also changes the font size of the pod title depending on whether the pod is full-width or not to enable proper formatting on the homepage.