Require due date when setting assignments and tests
The Assignment Schedule didn't seem to disable the button on an invalid
due date or show the warning, so this fixes the logic in Set Assignments
and copies it over too.
Quizzes have a slightly different check to assignments because the
deadlines work differently, but this may no longer be necessary in the
future.