Remove unused augmentation of questions results
The URL returned by the API does not contain this string except during
local development, and so this "augmentation" does nothing. Since we
do not use the URL anyhow, it is also unnecessary.
Passing through the action's results object revealed a mistake in the
typings; the type was really a SearchResultsWrapper all along.