Ensure coordinate questions use correct dimensions
The "default" item had an empty coordinates array in it; if it was not
modified by the user, it was treated as a zero dimensional coordinate by
the validator and misleading feedback was provided.
The number of dimensions can't be known outside the question, so some
refactoring was required. The spread syntax to clone the (mutable!)
default item was quite subtle, so I moved it to a function returning a
new item every time.
Also consistently default to 2 not 1 dimensions if the value is
undefined (this should not be possible, but is now consistent with the
code in the API).